accomodating to changes in rabbitmq structure

+++ src/main/java/com/linkedin/replica/wall/models/Reply.java
@@ -13,13 +13,13 @@ public class Reply {
     private String parentPostId;
     private String parentCommentId;
     private ArrayList<String> mentions;
-    private Long likesCount;
+    private int likesCount;
     private String text;
     private Date timestamp;
     private ArrayList<String> images;
     private ArrayList <String> urls;
 
-    public Reply(String replyId, String authorId, String parentPostId, String parentCommentId, ArrayList<String> mentions, Long likesCount, String text, Date timestamp, ArrayList<String> images,  ArrayList<String> urls ) {
+    public Reply(String replyId, String authorId, String parentPostId, String parentCommentId, ArrayList<String> mentions, int likesCount, String text, Date timestamp, ArrayList<String> images,  ArrayList<String> urls ) {
         this.replyId = replyId;
         this.authorId = authorId;
         this.parentPostId = parentPostId;
@@ -55,11 +55,11 @@ public class Reply {
         return mentions;
     }
 
-    public Long getLikesCount() {
+    public int getLikesCount() {
         return likesCount;
     }
 
-    public void setLikesCount(Long likesCount) {
+    public void setLikesCount(int likesCount) {
         this.likesCount = likesCount;
     }
